% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{IB}
\alias{IB}
\title{Spatial balance based on Moran's I index}
\usage{
IB(W, s)
}
\arguments{
\item{W}{a sparseMatrix, i.e, inheriting from \code{\link[Matrix]{sparseMatrix}}, representing the spatial weights. See \code{\link{wpik}}.}

\item{s}{a vector of size N with elements equal to 0 or 1. The value 1 indicates that the unit is selected while the value 0 is for non-chosen units.}
}
\value{
A numeric value that represent the spatial balance. It could be any real value between -1 (spread) and 1 (clustered).
}
\description{
This function implements the spatial balance based on Moran's I index.
}
\details{
This measure of the spatial balance is developped by Tillé et al. (2018).
It uses a corrected version of the traditional Moran's I index.
This estimator use spatial weights \eqn{w_{ij}} that indicates how a unit
\eqn{i} is far from the unit \eqn{j}.
 Such matrix is supposed to include inclusion probabilities in its computation, hence, the spatial weights matrix
\eqn{\bf W} is generally not symmetric. The spatial balance measure is equal to


\deqn{I_B =\frac{( \bf s- \bar{s}_w)^\top  W ( s- \bar{s}_w)}{\bf \sqrt{( s- \bar{s}_w)^\top  D ( s- \bar{s}_w) ( s- \bar{s}_w)^\top  B ( s- \bar{s}_w)}}}

where \eqn{\bf D} is the diagonal matrix containing the \eqn{w_i}, 

\deqn{ \bf \bar{s}_w =  1 \frac{ s^\top  W  1}{ 1^\top  W  1}}

and 

\deqn{ \bf B =  W^\top  D^{-1}  W - \frac{ W^\top  1 1^\top  W}{1^\top  W  1}}.

To specifiy the spatial weights uses the argument \code{W}. See \code{\link{wpik}}.
}
\examples{
\dontrun{
  N <- 100
  n <- 20
  x <- seq(1,sqrt(N),1)
  X <- expand.grid(x,x)
  pik <- rep(n/N,N)
  W <- wpik(as.matrix(X),pik,bound = 1,tore = TRUE,jitter = FALSE,toreBound = sqrt(N))
  W <- W - diag(diag(W))
  s <- wave(as.matrix(X),pik,tore = TRUE,jitter = TRUE,comment = TRUE)
  I1 <- IB(W,s)

  plot(X)
  points(X[s == 1,],pch = 16)
}

}
\references{
Tillé, Y., Dickson, M.M., Espa, G., and Guiliani, D. (2018). Measuring the spatial balance of a sample: A new measure based on Moran's I index.
\emph{Spatial Statistics}, 23, 182-192. \url{https://doi.org/10.1016/j.spasta.2018.02.001}
}
\seealso{
\code{\link{wpik}}
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
