% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IBarma}
\alias{IBarma}
\title{Spatial balance based on Moran's I index}
\usage{
IBarma(W, s)
}
\arguments{
\item{W}{A matrix representing the spatial weights indicating how close is the jth unit to the ith.}

\item{s}{A numeric vector of 0 and 1 that represent the sample.}
}
\value{
A numeric value that represent the spatial spreading measure.
}
\description{
This function implements the method from
}
