% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{IPFP}
\alias{IPFP}
\title{Iterative proportional fitting procedure (IPFP) or raking ratio.}
\usage{
IPFP(A, bh, bi, maxiter = 300L, tol = 1e-09, comment = FALSE)
}
\arguments{
\item{A}{a sparseMatrix, i.e, inheriting from \code{\link[Matrix]{sparseMatrix}}.}

\item{bh}{a vector representing the row margin.}

\item{bi}{a vector representing the columm margin.}

\item{maxiter}{an integer indicating the maximum iterations allowed.}

\item{tol}{a real value that the error should reach before stopped.}

\item{comment}{an optional logical value, indicating some informations during the execution. Default is \code{FALSE}.}
}
\value{
sparse matrix \code{A} adjusted.
}
\description{
Iterative proportional fitting procedure (IPFP) implemented for sparse Matrix.
}
\examples{
\dontrun{
X <- as.matrix(cbind(runif(1000),runif(1000)))
  pik <- rep(1/5,1000)
  A <- wpik(X,pik,bound = 1,tore = FALSE,jitter = FALSE,toreBound = 0.0)
  A <- A+t(A)
  isSymmetric(A)
  
  A1 <- IPFP(A,bh = rep(1,nrow(A)),
            bi = rep(1,nrow(A)),
            maxiter = 1000,
            tol = 1e-4,
            comment = TRUE)
  A2 <- IPFPsym(A,bh = rep(1,nrow(A)),
            bi = rep(1,nrow(A)),
            maxiter = 4000,
            tol = 1e-4,
            comment = TRUE) # takes more iterations because the symmetric structure is kept.
  rowSums(A1)
  rowSums(A2)
  colSums(A1)
  colSums(A2)
  
  isSymmetric(A1) # not symmetric
  isSymmetric(A2) # symmetric
}

}
\references{
Tillé, Y. (2019). \emph{Théorie des sondages : Échantillonage et estimation en populations finies} (2e édition ed.). Paris: Dunod
}
\seealso{
\code{\link{IPFPsym}}, \code{\link{wpik}}
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
