% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{projOp}
\alias{projOp}
\title{Projection operator}
\usage{
projOp(v, u)
}
\arguments{
\item{v}{vector that is projected.}

\item{u}{vector on which we project.}
}
\value{
The projection of the vector v onto the line spanned by the vector u.
}
\description{
This operator projects the vector v orthogonally onto the line spanned by vector u.
}
