% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowSumsiter}
\alias{rowSumsiter}
\title{Row sums on sparse matrix.}
\usage{
rowSumsiter(x)
}
\arguments{
\item{x}{A sparse matrix, i.e., inheriting from \code{\link[Matrix]{sparseMatrix}}.}
}
\value{
row sums of x.
}
\description{
Form row sums for sparseMatrix.
}
\details{
This function is designed to be used for internal \code{RcppArmadillo} functions. Nevertheless it could be applied in R.
It loops on the non-zero entries of the \code{\link[Matrix]{sparseMatrix}}. For general uses, the function \code{\link[Matrix]{rowSums}} should
be prefered.
}
\examples{
\dontrun{
sp <- rsparsematrix(5000,5000,density = 0.4)
system.time(test1 <- Matrix::colSums(sp))
system.time(test2 <- colSumsiter(sp))
}

}
\seealso{
\code{\link[Matrix]{colSums}}, \code{\link[Matrix]{rowSums}}.
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
