% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{STATENT2017}
\alias{STATENT2017}
\title{Structural statistics of Swiss enterprises}
\format{A data.frame with 9657 rows and 29 variables:
\describe{
  \item{E_KOORD}{coordinates E according to MN95.}
  \item{N_KOORD}{coordinates N according to MN95.}
  \item{X_KOORD}{coordinates x according to MN03.}
  \item{Y_KOORD}{coordinates y according to MN03.}
  \item{RELI}{...}
  \item{B1708T}{Total establishement.}
  \item{B1708S1}{establishement of sector 1.}
  \item{B1708S2}{establishement of sector 2.}
  \item{B1708S3}{establishement of sector 3.}
  \item{B1708EMPT}{Total employement.}
  \item{B1708EMPTS1}{employement of sector 1.}
  \item{B1708EMPFS1}{women employee of sector 1.}
  \item{B1708EMPMS1}{men employee of sector 1.}
  \item{B1708EMPTS2}{employemment of sector 2.}
  \item{B1708EMPFS2}{women employee of sector 2.}
  \item{B1708EMPMS2}{men employee o sector 2.}
  \item{B1708EMPTS3}{employement of sector 3.}
  \item{B1708EMPFS3}{women employee of sector 3.}   
  \item{B1708EMPMS3}{men employee of sector 3.}
  \item{B1708VZAT}{Total equivalent full time employement.}
  \item{B1708VZATS1}{equivalent full time employee of sector 1.}
  \item{B1708VZAFS1}{equivalent full time women employee of sector 1.}
  \item{B1708VZAMS1}{equivalent full time men employee of sector 1.}
  \item{B1708VZATS2}{equivalent full time employee of sector 2.}
  \item{B1708VZAFS2}{equivalent full time women employee of sector 2.}
  \item{B1708VZAMS2}{equivalent full time men employee of sector 2.}
  \item{B1708VZATS3}{equivalent full time employee of sector 3.}
  \item{B1708VZAFS3}{equivalent full time women employee of sector 3.}
  \item{B1708VZAMS3}{equivalent full time men employee of sector 3.}
}}
\source{
\url{https://www.bfs.admin.ch/bfs/fr/home/services/geostat/geodonnees-statistique-federale/etablissements-emplois/statistique-structurel-entreprises-statent-depuis-2011.html}
}
\usage{
STATENT2017
}
\description{
Only Swiss enterprises that have a total employement between 50 and 100.
}
\keyword{datasets}
