% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SwissCanton}
\alias{SwissCanton}
\title{Swiss canton boundaries}
\format{A data frame with 26 rows and 21 variables
\itemize{
  \item{KTNR}{Anonymized Mechanical Turk Worker ID}
  \item{KTNAME}{Anonymized Mechanical Turk Worker ID}
  \item{GRNR}{Anonymized Mechanical Turk Worker ID}
  \item{AREA_HA}{Anonymized Mechanical Turk Worker ID}
  \item{X_MIN}{minimum x coordinate of the territory [m] according to MN03.}
  \item{X_MAX}{maximum x coordinate of the territory [m] according to MN03.}
  \item{Y_MIN}{minimum x coordinate of the territory [m] according to MN03.}
  \item{Y_MAX}{maximum x coordinate of the territory [m] according to MN03.}
  \item{X_CNTR}{hectometric coordinate x of the central coordinates according to MN03. }
  \item{Y_CNTR}{hectometric coordinate y of the central coordinates according to MN03.}
  \item{Z_MIN}{minimum elevation of the territory}
  \item{Z_MAX}{maximum elevation of the territory}
  \item{Z_AVG}{avergae elecation of the territory}
  \item{Z_MED}{median elevation of the territory.}
  \item{E_MIN}{minimum E coordinate of the territory [m] according to MN95.}
  \item{E_MAX}{maximum E coordinate of the territory [m] according to MN95.}
  \item{N_MIN}{minimum N coordinate of the territory [m] according to MN95.}
  \item{N_MAX}{maximum N coordinate of the territory [m] according to MN95.}
  \item{E_CNTR}{hectometric coordinate E of the central coordinates according to MN95.}
  \item{N_CNTR}{hectometric coordinate N of the central coordinates according to MN95.}
  \item{geometry}{Trial number, from 1..NNN}
}}
\usage{
SwissCanton
}
\description{
This is data from the first experiment ever to try XYZ using Mechanical
Turk workers.
}
\keyword{datasets}
