% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{SwissCommune}
\alias{SwissCommune}
\title{Swiss commune boundaries}
\format{A \code{\link[sf]{sf}} object with 2215 rows and 26 variables:
\describe{
  \item{GMDHISTID}{Identification number of the commune.}
  \item{GMDNR}{Commune number.}
  \item{GMDNAME}{Commune name.}
  \item{BZHISTID}{Identifaction number of the district.}
  \item{BZNR}{District number.}
  \item{KTNR}{Canton number.}
  \item{GRNR}{Region number.}
  \item{AREA_HA}{Area of the considered territory.}
  \item{E_MIN}{minimum E coordinate of the territory [m] according to MN95.}
  \item{E_MAX}{maximum E coordinate of the territory [m] according to MN95.}
  \item{N_MIN}{minimum N coordinate of the territory [m] according to MN95.}
  \item{N_MAX}{maximum N coordinate of the territory [m] according to MN95.}
  \item{E_CNTR}{hectometric coordinate E of the central coordinates according to MN95.}
  \item{N_CNTR}{hectometric coordinate N of the central coordinates according to MN95.}  
  \item{Z_MIN}{minimum elevation of the territory}
  \item{Z_MAX}{maximum elevation of the territory}
  \item{Z_AVG}{avergae elecation of the territory}
  \item{Z_MED}{median elevation of the territory.}
  \item{Z_CNTR}{median elevation of the territory.}
  \item{X_MIN}{minimum x coordinate of the territory [m] according to MN03.}
  \item{X_MAX}{maximum x coordinate of the territory [m] according to MN03.}
  \item{Y_MIN}{minimum x coordinate of the territory [m] according to MN03.}
  \item{Y_MAX}{maximum x coordinate of the territory [m] according to MN03.}
  \item{X_CNTR}{hectometric coordinate x of the central coordinates according to MN03. }
  \item{Y_CNTR}{hectometric coordinate y of the central coordinates according to MN03.}
  \item{geometry}{geometry considered}
}}
\source{
\url{https://www.bfs.admin.ch/bfs/fr/home/services/geostat/geodonnees-statistique-federale/limites-administratives/limites-communales-generalisees.html}
}
\usage{
SwissCommune
}
\description{
\code{\link[sf]{sf}} sf object that represent the boundaries of the Switzerland.
}
\keyword{datasets}
