% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{SwissLake}
\alias{SwissLake}
\title{Swiss lakes boundaries}
\format{A \code{\link[sf]{sf}} object with 23 rows and 16 variables:
\describe{
  \item{GMDNR}{...}
  \item{GMDNAME}{...}
  \item{SEE_HA}{Anonymized Mechanical Turk Worker ID}
  \item{X_MIN}{minimum x coordinate of the territory [m] according to MN03.}
  \item{X_MAX}{maximum x coordinate of the territory [m] according to MN03.}
  \item{Y_MIN}{minimum x coordinate of the territory [m] according to MN03.}
  \item{Y_MAX}{maximum x coordinate of the territory [m] according to MN03.}
  \item{X_CNTR}{hectometric coordinate x of the central coordinates according to MN03. }
  \item{Y_CNTR}{hectometric coordinate y of the central coordinates according to MN03.}
  \item{E_MIN}{minimum E coordinate of the territory [m] according to MN95.}
  \item{E_MAX}{maximum E coordinate of the territory [m] according to MN95.}
  \item{N_MIN}{minimum N coordinate of the territory [m] according to MN95.}
  \item{N_MAX}{maximum N coordinate of the territory [m] according to MN95.}
  \item{E_CNTR}{hectometric coordinate E of the central coordinates according to MN95.}
  \item{N_CNTR}{hectometric coordinate N of the central coordinates according to MN95.}  
  \item{geometry}{Trial number, from 1..NNN}
}}
\source{
\url{https://www.bfs.admin.ch/bfs/fr/home/services/geostat/geodonnees-statistique-federale/limites-administratives/limites-communales-generalisees.html}
}
\usage{
SwissLake
}
\description{
\code{\link[sf]{sf}} sf object that represent the boundaries of the Switzerland.
}
\keyword{datasets}
