% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varHR.R
\name{varHR}
\alias{varHR}
\title{Variance estimated by CP}
\usage{
varHR(y, pik, s)
}
\arguments{
\item{y}{vector of size n that represent the variable of interest.}

\item{pik}{vector of the inclusion probabilites. The length should be equal to n.}

\item{s}{vector of size n with elements equal to the index of the selected units.}
}
\value{
variance
}
\description{
Variance estimated by CP
}
