% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varIRS.R
\name{varIRS}
\alias{varIRS}
\title{Variance estimated by independant random sampling}
\usage{
varIRS(y, pik, s)
}
\arguments{
\item{y}{vector of size N that represent the variable of interest.}

\item{pik}{vector of the inclusion probabilites. The length should be equal to N.}

\item{s}{vector of size N with elements equal 0 or 1. The value 1 indicates that the unit is selected while the value 0 is for non-chosen unit.}
}
\value{
variance
}
\description{
Variance estimated by independant random sampling
}
\examples{
\dontrun{

data("STATENT2017")
pik <- sampling::inclusionprobabilities(STATENT2017$B1708T,10)
y <- STATENT2017$B1708EMPT
X <- as.matrix(scale(STATENT2017[,1:2]))
s <- rep(0,length(pik))
 s[BalancedSampling::lpm1(pik,X)] <- 1
varIRS(y,pik,s)

}

}
