% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varNBH.R
\name{varNBH}
\alias{varNBH}
\title{Variance estimated by NBH}
\usage{
varNBH(y, pik, s, X, d = 4)
}
\arguments{
\item{y}{vector of size n that represent the variable of interest.}

\item{pik}{vector of the inclusion probabilites. The length should be equal to n.}

\item{s}{vector of size n with elements equal to the index of the selected units.}

\item{X}{coordinates of the sampled units.}

\item{d}{number of neighbors}
}
\value{
variance
}
\description{
Variance estimated by NBH
}
\examples{
\dontrun{
y <- y[which(s==1)]
pik <- pik[which(s ==1)]
X <- X[which(s == 1),]
s <- which(s == 1)
d = 4
}

}
