% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{varNBH2}
\alias{varNBH2}
\title{Local mean variance estimator}
\usage{
varNBH2(y, pik, s, X, d)
}
\arguments{
\item{y}{vector of size n that represent the variable of interest.}

\item{pik}{vector of the inclusion probabilites. The length should be equal to n.}

\item{s}{vector of size n with elements equal to the index of the selected units.}

\item{X}{coordinates of the sampled units.}

\item{d}{number of neighbors}
}
\value{
An integer the variance estimator
}
\description{
Calculates the varaiance estimator based on the local mean.
}
\details{
blablalb
}
\references{
Stevens, D. L. Jr. and Olsen, A. R. (2004). Spatially balanced sampling of natural resources.
\emph{Journal of the American Statistical Association 99, 262-278}
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
