% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{wave}
\alias{wave}
\title{Weakly associated vectors sampling}
\usage{
wave(X, pik, bound = 1, tore = FALSE, jitter = FALSE, oneD = FALSE,
  comment = FALSE)
}
\arguments{
\item{X}{matrix of size N x 2 representing the spatial coordinates.}

\item{pik}{vector of the inclusion probabilites. The length should be equal to N.}

\item{bound}{a scalar representing the bound to reach. See Details. Default is 1.}

\item{tore}{an optional logical value, if we are considering the distance on a tore. See Details. Default is \code{TRUE}.}

\item{jitter}{an optional logical value, if you would use a jitter perturbation. See Details. Default is \code{FALSE}.}

\item{oneD}{an optional logical value, specifying if we are in one dimension. Default is \code{FALSE}.}

\item{comment}{an optional logical value, indicating some informations during the execution. Default is \code{FALSE}.}
}
\value{
A vector of size N with elements equal 0 or 1. The value 1 indicates that the unit is selected while the value 0 is for non-chosen unit.
}
\description{
Select a spread spatial samples from inclusion probabilities using the weakly associated vectors sampling method.
}
\details{
Weakly associated vectors sampling find the vector \eqn{\bf u} that is
the weakest associated vector of the linear transformations generated by a contiguity table \eqn{ \bf W}.
The matrix \eqn{\bf W} is calculated from the matrix of coordinates \eqn{\bf X} by the function \code{\link{wpik}}.
The vector \eqn{\bf u} is calculated from a singular value decompositon or a QR decomposition depending on the rank of the matrix 
\eqn{\bf W}.

The procedure then update the inclusion probabilities vector \eqn{\bf\pi} and the matrix \eqn{\bf W} following the cube 
method developped by Deville and Tillé (2004), improved by Chauvet and Tillé (2006) and 
used in spatial sampling by Grafström and Tillé (2013).

For more informations on the options \code{tore} and \code{toreBound}, see \code{\link{distUnitk}}.

For more informations on the option \code{jitter}, see \code{\link{wpik}}.
}
\examples{
\dontrun{
  N <- 36 # 6 x 6 grid
  n <- 12 # number of unit selected
  x <- seq(1,sqrt(N),1)
  X <- as.matrix(cbind(rep(x,times = sqrt(N)),rep(x,each = sqrt(N))))
  pik <- rep(n/N,N)
  s <- wave(X,pik, tore = TRUE,jitter = FALSE)
  plot(X)
  points(X[s == 1,],pch = 16)
}



}
\references{
Deville, J. C. and Tillé, Y. (2004). Efficient balanced sampling: the cube method. Biometrika, 91(4), 893-912

Chauvet, G. and Tillé, Y. (2006). A fast algorithm for balanced sampling. Computational Statistics, 21(1), 53-62

Grafström, A. and Tillé, Y. (2013). Doubly balanced spatial sampling with spreading and restitution of auxiliary totals. Environmetrics, 24(2), 120-131
}
\seealso{
\code{\link{wpik}}, \code{\link{distUnitk}}.
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
