% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveR.R
\name{waveR}
\alias{waveR}
\title{Strongly spread cube method}
\usage{
waveR(X, pik, tore = TRUE, jitter = FALSE)
}
\arguments{
\item{X}{A matrix of size N x 2, it should represent the 2D position of the units.}

\item{pik}{vector of inclusion probabilites.}

\item{tore}{logical, if we are considering the distance on a tore. Default is TRUE.}

\item{jitter}{logical, if you would use a jitter perturbation. See Details for more infomrations. Default is FALSE.}
}
\value{
A numeric vector (with elements 0 and 1) of size N, the population size. 1 if it is in the sample and 0 otherwise.
}
\description{
Find a strongly spread sample of a given population of size N.
}
\examples{
\dontrun{



#### Example with 6x6 grid (36 units)

N <- 6
n <- 18
x <- seq(1,N,1)
y <- seq(1,N,1)
X <- as.matrix(expand.grid(x,y))
pik <- rep(n/(N*N),N*N)
s <- waveR(as.matrix(X),pik,tore = TRUE,jitter = FALSE)
plot(X)
points(X[s == 1,],pch = 16)
}
}
