% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{wpik2}
\alias{wpik2}
\title{Spatial weights from pik}
\usage{
wpik2(X, pik, tore, jitter, toreBound)
}
\arguments{
\item{X}{matrix of size N x 2 representing the spatial coordinates.}

\item{pik}{vector of the inclusion probabilites. The length should be equal to N.}

\item{tore}{an optional logical value, if we are considering the distance on a tore. See \code{\link{distUnitk}}. Default is TRUE.}

\item{jitter}{an optional logical value, if you would use a jitter perturbation. See Details for more infomrations. Default is FALSE.}

\item{toreBound}{A numeric value that specify the size of the grid.}
}
\value{
A sparse matrix representing the spatial weights.
}
\description{
Spatial weights calculated from inclusion probabilies taking distance between units into account. It is a direct
implementation of the spatial weights specified in [Tillé et al., 2018].
}
\details{
Spatial weights indicates how close the units are frome each others. Hence a large value \eqn{w_{ij}} means that the unit \eqn{i} 
is close to the unit \eqn{j}. This function consider that if \eqn{i} were selected in the sample drawn from the population then
\eqn{i} would represent \eqn{1/\pi_i} units in the population and, as a consequence, it would onl be natural to consider
that \eqn{i} has \eqn{k_i = (1/\pi_i -1 )} neighbours in the population. The \eqn{k_i} neighbours can be the nearest 
neighbours of \eqn{i} according to the distance. The weights are so calculated as follows,

\deqn{ w_{ij} = 1}
 if unit \eqn{j \in N_{\lfloor k_i \rfloor}}.
\deqn{ w_{ij} = k_i - \lfloor k_i \rfloor}
 if unit j is the \eqn{\lceil k_i \rceil} the nearest neighbour of i.
\deqn{w_{ij} = 0}
 otherwise.

where \eqn{ \lfloor k_i \rfloor} and \eqn{\lceil k_i \rceil} bet he inferior and the superior integers of \eqn{k_i}.
}
\references{
Tillé, Y., Dickson, M.M., Espa, G., and Guiliani, D. (2018). Measuring the spatial balance of a sample: A new measure based on Moran's I index.
\emph{Spatial Statistics}, 23, 182-192. \url{https://doi.org/10.1016/j.spasta.2018.02.001}
}
\seealso{
\code{\link{wpik2}}, \code{\link{distUnitk}}, \code{\link{wave}}.
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
