% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colSumsiter}
\alias{colSumsiter}
\title{column sum on sparse matrix}
\usage{
colSumsiter(x)
}
\arguments{
\item{x}{A sparse matrix}
}
\value{
column sum of the sparse matrix
}
\description{
column sum that iterate only on the non-zero entries of the matrix.
}
