% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colSumsiter}
\alias{colSumsiter}
\title{Column sums on sparse matrix.}
\usage{
colSumsiter(x)
}
\arguments{
\item{x}{A sparse matrix of class \code{\link[Matrix]{sparseMatrix}}.}
}
\value{
A numeric vector that represent the sums of the column of x.
}
\description{
Same usage of the function \code{\link[Matrix]{colSums}}.
}
\details{
This function is designed to be used for other \code{RcppArmadillo} function inside of the package. Nevertheless it could be used in R.
It only loops on the non-zero entries of the sparseMatrix. The function \code{\link[Matrix]{colSums}} should be prefer because is 
a little bit faster.
}
\examples{
\dontrun{
sp <- rsparsematrix(5000,5000,density = 0.4)
system.time(test1 <- Matrix::colSums(sp))
system.time(test2 <- colSumsiter(sp))
}
}
\seealso{
\code{\link[Matrix]{colSums}}, \code{\link[Matrix]{rowSums}}.
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
