% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distUnitk}
\alias{distUnitk}
\title{Square of the euclidean distance of the unit k.}
\usage{
distUnitk(X, k, tore, toreBound)
}
\arguments{
\item{X}{matrix of size N x 2 representing the spatial coordinates.}

\item{k}{the unit index to be used.}

\item{tore}{an optional logical value, if we are considering the distance on a tore. See Details.}

\item{toreBound}{an optional numeric value that specify the size of the grid.}
}
\value{
distance values of the unit k.
}
\description{
Calculate the squared euclidean distance from the unit k to the other units.
}
\details{
Let \eqn{x_k = (x_{k_1},x_{k_2})} be the spatial coordinates of the unit \eqn{k}. The classical euclidean distance is given by

\deqn{d^2(k,l) = (x_k - x_l)^\top (x_k - x_l). }

When the points are distributed on a \eqn{r_1 \times r_2} regular grid of \eqn{R^2}.
It is possible to consider the units like they were placed on a tore. Specifically,
we could consider two units on the same column (resp. row) that are on the opposite have a small distance,

\deqn{ d(k,l) = min( (x_{k_1} - x_{l_1})^2,
                      (r_1 + x_{k_1} - x_{l_1})^2,
                      (r_1 - x_{k_1} + x_{l_1})^2) +}
\deqn{ min( (x_{k_2} - x_{l_2})^2,
                      (r_2 + x_{k_2} - x_{l_2})^2,
                      (r_2 - x_{k_2} + x_{l_2})^2)}.

The option \code{toreBound} specify the size of the grid in the case of \eqn{r_1 = r_2 = r}. 
It is omitted if the tore option is equal to FALSE.
}
\examples{
\dontrun{
  N <- 5
  x <- seq(1,N,1)
  X <- as.matrix(expand.grid(x,x))
  distUnitk(X,k = 2,tore = TRUE,toreBound = 5)
  distUnitk(X,k = 2,tore = FALSE,toreBound = -1)
}
}
\seealso{
\code{\link{wpik}}, \code{\link{wave}} and \code{\link[stats]{dist}}.
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
