% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distUnitk}
\alias{distUnitk}
\title{Euclidean distance with tore option}
\usage{
distUnitk(X, k, tore, toreBound)
}
\arguments{
\item{X}{A matrix of size N x 2, it should represent the 2D position of the units.}

\item{k}{An integer value, the index of the unit.}

\item{tore}{logical, if we are considering the distance on a tore. Default is TRUE.}

\item{toreBound}{A numeric value that specify the size of the grid.}
}
\value{
A vector that represent the distance values.
}
\description{
Calculate the distance from the unit k to the other units.
}
