% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{IB}
\alias{IB}
\title{Spatial balance based on Moran's I index}
\usage{
IB(W, s)
}
\arguments{
\item{W}{A \code{\link[Matrix]{sparseMatrix}} representing the spatial weights. See \code{\link{wpik}}.}

\item{s}{A numeric vector of 0 and 1 that represent the sample.}
}
\value{
A numeric value that represent the spatial spreading measure. It could be any real value between -1 (spread) and 1 (clustered).
}
\description{
This function implements the spatial balanced based on Moran's I index.
}
\examples{
\dontrun{
  N <- 100
  n <- 20
  x <- seq(1,sqrt(N),1)
  X <- expand.grid(x,x)
  pik <- rep(n/N,N)
  W <- wpik(as.matrix(X),pik,bound = 1,tore = TRUE,jitter = FALSE,toreBound = sqrt(N))
  W <- W - diag(diag(W))
  s <- round(wave(as.matrix(X),pik,tore = TRUE,jitter = TRUE,comment = TRUE))
  system.time(I1 <- IB(W,s))

  plot(X)
  points(X[s == 1,],pch = 16)
}

}
\references{
Tillé, Y., Dickson, M.M., Espa, G., and Guiliani, D. (2018). Measuring the spatial balance of a sample: A new measure based on Moran's I index.
\emph{Spatial Statistics}, 23, 182-192. \url{https://doi.org/10.1016/j.spasta.2018.02.001}
}
\seealso{
\code{\link{wpik}}
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
