% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{projOp}
\alias{projOp}
\title{Projection operator}
\usage{
projOp(v, u)
}
\arguments{
\item{v}{vector projected.}

\item{u}{vector that define the line on which we project.}
}
\value{
The projection of the vector v onto the line spanned by the vector u.
}
\description{
This operator projects the vector v orthogonally onto the line spanned by vector u.
}
\details{
The projection operator is defined by :

\deqn{proj_u(v) = \frac{\langle u , v \rangle}{\langle u, u \rangle} u}
 where \eqn{\langle . , . \rangle} is the inner product also written \eqn{u^\top v}.
}
\examples{
\dontrun{
u = c(0,1)
v = c(1,1)
projOp(v,u)
v - projOp(v,u)
}

}
\references{
\url{https://en.wikipedia.org/wiki/Projection_(linear_algebra)}s
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
