% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowSumsiter}
\alias{rowSumsiter}
\title{row sum on sparse matrix}
\usage{
rowSumsiter(x)
}
\arguments{
\item{x}{A sparse matrix}
}
\value{
row sum of the sparse matrix
}
\description{
row sum that iterate only on the non-zero entries of the matrix.
}
