% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wave}
\alias{wave}
\title{spreadcube implemented in RcppArmadillo}
\usage{
wave(X, pik, tore = TRUE, jitter = TRUE, oneD = FALSE,
  comment = FALSE)
}
\arguments{
\item{X}{A matrix of size N x 2, it should represent the 2D position of the units.}

\item{pik}{vector of inclusion probabilites.}

\item{tore}{logical, if we are considering the distance on a tore. Default is TRUE.}

\item{jitter}{logical, if you would use a jitter perturbation. See Details for more infomrations. Default is FALSE.}

\item{bound}{bound to reached before a new strata is considered. Default is 1.}

\item{toreBound}{A numeric value that specify the size of the grid.}
}
\value{
A sparse matrix.
}
\description{
Same as spreadcube but with RcppArmadillo implementation.
}
