% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{wave}
\alias{wave}
\title{Weakly associated vectors sampling}
\usage{
wave(X, pik, bound = 1, tore = FALSE, jitter = FALSE, oneD = FALSE,
  comment = FALSE)
}
\arguments{
\item{X}{matrix of size N x 2 representing the spatial position.}

\item{pik}{vector of the inclusion probabilites. The length should be equal to N.}

\item{bound}{A scalar representing the bound to reach before a new strata is considered. See \code{\link{wpik}}. Default is 1.}

\item{tore}{an optional logical value, if we are considering the distance on a tore. See \code{\link{distUnitk}}. Default is TRUE.}

\item{jitter}{an optional logical value, if you would use a jitter perturbation. See Details for more infomrations. Default is FALSE.}

\item{oneD}{an optional logical value, specifying if we are in one dimension. Default is FALSE.}

\item{comment}{an optional logical value, indicating some informations during the execution. Default is FALSE.}
}
\value{
A vector of size N with elements equal 0 or 1. The value 1 indicates that the unit is selected while the value 0 is for non-selected unit.
}
\description{
Select spread spatial samples with inclusion probabilities using the weakly associated vectors sampling method.
}
\details{
A contiguity table \eqn{W} is calculated from the matrix \eqn{X}.
\code{wave} sampling estimate a vector \eqn{u} that is
the weakest associated vector of the linear transformations generated by \eqn{W}.
The vector \eqn{u} is then used to update the inclusion probabilities vector \eqn{\pi_k}.
}
\examples{
\dontrun{
  N <- 36 # 6 x 6 grid
  n <- 12 # number of unit selected
  x <- seq(1,sqrt(N),1)
  X <- as.matrix(cbind(rep(x,times = sqrt(N)),rep(x,each = sqrt(N))))
  pik <- rep(n/N,N)
  s <- wave(X,pik, tore = TRUE,jitter = FALSE)
  plot(X)
  points(X[s == 1,],pch = 16)
}



}
\references{
\url{https://en.wikipedia.org/wiki/Singular_value_decomposition}
}
\seealso{
\code{\link{wpik}}, \code{\link{distUnitk}}.
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
