% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{wpik}
\alias{wpik}
\title{Spatial weights from pik}
\usage{
wpik(X, pik, bound = 1, tore = FALSE, jitter = FALSE,
  toreBound = -1)
}
\arguments{
\item{X}{matrix of size N x 2 representing the spatial coordinates.}

\item{pik}{vector of the inclusion probabilites. The length should be equal to N.}

\item{bound}{A scalar representing the bound to reach before a new strata is considered. See \code{\link{wpik}}. Default is 1.}

\item{tore}{an optional logical value, if we are considering the distance on a tore. See \code{\link{distUnitk}}. Default is TRUE.}

\item{jitter}{an optional logical value, if you would use a jitter perturbation. See Details for more infomrations. Default is FALSE.}

\item{toreBound}{A numeric value that specify the size of the grid.}
}
\value{
A sparse matrix representing the spatial weights.
}
\description{
Spatial weights calculated from inclusion probabilies taking distance between units into account.
}
\details{
Spatial weights indicates how close the units are frome each others. Hence a large value \eqn{w_{ij}} means that the unit i 
is close to the unit j. This function consider that a unit represents its neighbor till their inclusion probabilities
sum to 1.

We define \eqn{H_i} the set of the nearest neighbor of the unit i including i such that the sum of their inclusion
probabilities is just greater than 1. Moreover, let \eqn{h_i = card{H_i}}, the number of elements in \eqn{H_i}.
The matrix \eqn{W} is then defined as follows,

\deqn{ w_{ij} = \pi_j}
 if unit j is in the set of the  \eqn{h_i - 1} nearest neighbor of \eqn{i}.
\deqn{ w_{ij} = g_j}
 if unit j is the \eqn{h_i} nearest neighbour of \eqn{i}.
\deqn{w_{ij} = 0}
 otherwise.

where \eqn{g_j = 1- (\sum_{k \in H_i} \pi_k -\pi_j)}. Hence, the ith row of the matrix represents
neighborhood or stratum of the unit such that the inclusion probabilities sum up to 1 and
the ith column the weights that unit i takes for each stratum.
}
\references{
Tillé, Y., Dickson, M.M., Espa, G., and Guiliani, D. (2018). Measuring the spatial balance of a sample: A new measure based on Moran's I index.
\emph{Spatial Statistics}, 23, 182-192. \url{https://doi.org/10.1016/j.spasta.2018.02.001}
}
\seealso{
\code{\link{wpik2}}, \code{\link{distUnitk}}, \code{\link{wave}}.
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
