% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wpik}
\alias{wpik}
\title{Weights calculated from pik}
\usage{
wpik(X, pik, bound, tore, jitter, toreBound)
}
\arguments{
\item{X}{A matrix of size N x 2, it should represent the 2D position of the units.}

\item{pik}{vector of inclusion probabilites.}

\item{bound}{bound to reached before a new strata is considered. Default is 1.}

\item{tore}{logical, if we are considering the distance on a tore. Default is TRUE.}

\item{jitter}{logical, if you would use a jitter perturbation. See Details for more infomrations. Default is FALSE.}

\item{toreBound}{A numeric value that specify the size of the grid.}
}
\value{
A sparse matrix.
}
\description{
The weights are calculated in a way that a unit represents its neighbor till their inclusion probabilities sum to 1.
Hence each line represent a strata centered around the unit i and that sum up to 1.
}
