% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{wpikInv}
\alias{wpikInv}
\title{Spatial weights from inverse inclusion probabilities}
\usage{
wpikInv(X, pik, tore = FALSE, jitter = FALSE, toreBound = -1)
}
\arguments{
\item{X}{matrix of size N x 2 representing the spatial coordinates.}

\item{pik}{vector of the inclusion probabilites. The length should be equal to N.}

\item{tore}{an optional logical value, if we are considering the distance on a tore. Default is \code{FALSE}.}

\item{jitter}{an optional logical value, if you would use a jitter perturbation. See Details for more infomrations. Default is \code{FALSE}.}

\item{toreBound}{a numeric value that specify the size of the grid. Default is -1.}
}
\value{
A sparse matrix representing the spatial weights.
}
\description{
Spatial weights calculated from inclusion probabilies taking distance between units into account. It is a direct
implementation of the spatial weights specified in Tillé et al., (2018).
}
\details{
Spatial weights indicates how close the units are frome each others. Hence a large value \eqn{w_{ij}} means that the unit \eqn{i} 
is close to the unit \eqn{j}. This function consider that if \eqn{i} were selected in the sample drawn from the population then
\eqn{i} would represent \eqn{1/\pi_i} units in the population and, as a consequence, it would be natural to consider
that \eqn{i} has \eqn{k_i = (1/\pi_i -1 )} neighbours in the population. The \eqn{k_i} neighbours can be the nearest 
neighbours of \eqn{i} according to the distance. The weights are so calculated as follows,

\deqn{ w_{ij} = 1,}
 if unit \eqn{j \in N_{\lfloor k_i \rfloor}}.
\deqn{ w_{ij} = k_i - \lfloor k_i \rfloor,}
 if unit \eqn{j} is the \eqn{\lceil k_i \rceil} the nearest neighbour of \eqn{i}.
\deqn{w_{ij} = 0,}
 otherwise. \eqn{ \lfloor k_i \rfloor} and \eqn{\lceil k_i \rceil} be the inferior and the superior integers of \eqn{k_i}.

The option \code{jitter} will add a small normally distributed perturbation \code{rnorm(0,0.01)} to the coordinates
of the centroid of the stratum considered. This could be useful if there are many unit that have the same distances.
Indeed, if two units have the same distance and are the last unit before that the bound is reached, then the weights
of the both units is updated. If a jitter perturbation is used then all the distance are different and only one unit
weight is update such that the bound is reached. 

The jitter perturbation is generated at the beginning of the procedure such that each stratum is shifted by the same perturbation.
}
\examples{
\dontrun{
X <- cbind(runif(1000),runif(1000))
pik <- sampling::inclusionprobabilities(runif(1000),100)
d <- array(rep(0,1000*1000),c(1000,1000))
for(i in 1:1000){
  d[i,] <- distUnitk(X,k =i,tore = FALSE,toreBound = 0)
}
system.time(W <- wpikInv(X,pik = pik,tore = FALSE,jitter = FALSE,toreBound =0))
}

}
\references{
Tillé, Y., Dickson, M.M., Espa, G., and Guiliani, D. (2018). Measuring the spatial balance of a sample: A new measure based on Moran's I index.
\emph{Spatial Statistics}, 23, 182-192. \url{https://doi.org/10.1016/j.spasta.2018.02.001}
}
\seealso{
\code{\link{wpik}}, \code{\link{distUnitk}}, \code{\link{wave}}.
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
